chore: move confirmationDialog and Menu out of lab
