+++ solara/components/confirmation_dialog.py
