+++ solara/lab/components/__init__.py
@@ -1,8 +1,6 @@
 from .chat import ChatBox, ChatInput, ChatMessage
-from .confirmation_dialog import ConfirmationDialog
 from .input_date import InputDate, InputDateRange
 from .input_time import InputTime as InputTime
-from .menu import ClickMenu, ContextMenu, Menu
 from .tabs import Tab, Tabs
 from .theming import ThemeToggle, theme, use_dark_effective
 
@@ -11,13 +9,9 @@ __all__ = [
     "ChatBox",
     "ChatInput",
     "ChatMessage",
-    "ConfirmationDialog",
     "InputDate",
     "InputDateRange",
     "InputTime",
-    "ClickMenu",
-    "ContextMenu",
-    "Menu",
     "Tab",
     "Tabs",
     "ThemeToggle",
